+++ limber/operators/python_operator/gcp.py
@@ -1,14 +1,9 @@
 import os
-from constructs import Construct
 from limber.operators.python_operator.operator import PythonOperator
-from cdktf import App, TerraformStack, TerraformOutput, TerraformAsset, AssetType, Token
-from limber.imports.google import PubsubTopic
 from limber.imports.google import (
-    CloudfunctionsFunctionIamBinding,
-    GoogleProvider,
     CloudfunctionsFunction,
-    StorageBucket,
     StorageBucketObject,
+    PubsubTopic,
 )
