+++ limber/operators/python_operator/operator.py
@@ -130,50 +130,3 @@ class PythonOperator(Operator):
         message_bytes = message_json.encode('utf-8')
         publish_future = publisher.publish(topic_path, data=message_bytes)
         publish_future.result()
-
-    def get_terraform_json_old(self, *, folder) -> {}:
-
-        hash = self._write_cloud_function_code(folder=folder)
-
-        source_dir = f"{self.dag.dag_id}/{self.task_id}"
-
-        if len(self.upstream_tasks) > 0:
-            trigger_resource = f"task_{self.dag.dag_id}_{self.upstream_tasks[0]}"
-        else:
-            trigger_resource = f"dag_{self.dag.dag_id}"
-
-        configuration = {
-            "resource": {
-                "google_storage_bucket_object": {
-                    f"task_{self.task_id}": {
-                        "name": f"{source_dir}_{hash}.zip",
-                        "bucket": "${google_storage_bucket.bucket.name}",
-                        "source": f"{source_dir}.zip"
-                    }
-                },
-                "google_cloudfunctions_function": {
-                    f"function_{self.task_id}": {
-                        "name": f"{self.dag.dag_id}-{self.task_id}",
-                        "description": self.description,
-                        "runtime": "python37",
-                        "available_memory_mb": self.memory,
-                        "timeout": self.timeout,
-                        "service_account_email": os.environ["CLOUD_FUNCTIONS_SERVICE_ACCOUNT_EMAIL"],
-                        "source_archive_bucket": "${google_storage_bucket.bucket.name}",
-                        "source_archive_object": "${google_storage_bucket_object.task_"+self.task_id+".name}",
-                        "event_trigger": {
-                            "event_type": "providers/cloud.pubsub/eventTypes/topic.publish",
-                            "resource": trigger_resource
-                        },
-                        "entry_point": "cloudfunction_execution"
-                    }
-                },
-                "google_pubsub_topic": {
-                    f"task_{self.dag.dag_id}_{self.task_id}": {
-                        "name": f"task_{self.dag.dag_id}_{self.task_id}"
-                    }
-                }
-            }
-        }
-
-        return configuration
